+++ src/Dhall/Core.hs-boot
@@ -1,0 +1,7 @@
+module Dhall.Core where
+
+data Const
+
+data Var
+
+data Expr s a
