+++ src/Dhall/Pretty.hs
@@ -1,0 +1,3 @@
+module Dhall.Pretty (prettyExpr) where
+
+import Dhall.Pretty.Internal
